requirements.txt
streamlit
google-genai
python-docx
jsonschema
python-dotenv